Feed Page path changed to '/'
